modularized graph, possibly dead for showing graph
